requirements.txt
asgiref==3.8.1
cloudinary==1.36.0
crispy-bootstrap5==2024.2
dj3-cloudinary-storage==0.0.6
Django==4.2.13
django-allauth==0.62.1
django-autoslug==1.9.9
django-crispy-forms==2.1
django-richtextfield==1.6.2
django-summernote==0.8.20.0
sqlparse==0.5.0
urllib3==1.26.18
